simba/requirements.txt
fastkde>2
h5py==3.4.0
matplotlib==3.4.3
mpl_axes_aligner==1.3
munch==2.5.0
numpy==1.22.0
progressbar2==4.0.0
pyyaml==6.0
pyzmq==22.3.0
scipy==1.7.1
cheetah-accelerator>=0.7.5